requirements_vc2.txt
decord==0.6.0
einops==0.3.0
imageio==2.9.0
numpy==1.24.2
omegaconf==2.1.1
opencv_python==4.10.0.84
pandas==2.0.0
Pillow==9.5.0
pytorch_lightning==2.3.3
PyYAML==6.0
setuptools==65.6.3
tqdm==4.65.0
transformers==4.25.1
moviepy==1.0.3
av==12.3.0
xformers==0.0.28.post1
gradio==4.44.1
timm==1.0.13
scikit-learn==1.3.2
open_clip_torch==2.22.0
kornia==0.7.3